+++ src/main/java/com/vss_market/network/s2c/S2CPayload.java
@@ -49,7 +49,7 @@ public class S2CPayload {
         RPCPacketDistributor.rpcToPlayer(player, OPEN_MARKET_SCREEN, createPayload(player, selectedShopId, selectedListingId, view, shopListScroll, manageListingScroll, result));
     }
 
-    public static void openUpload(ServerPlayer player, MarketResult result, ItemStack stack, int price, int bundleSize, int stock, boolean purchaseOrder) {
+    public static void openUpload(ServerPlayer player, MarketResult result, ItemStack stack, double price, int bundleSize, int stock, boolean purchaseOrder) {
         var payload = createPayload(player, player.getUUID().toString(), "", "UPLOAD", 0f, 0f, result);
         payload.setUploadStack(stack)
                 .setUploadPrice(price)
